+++ packages/capabilities/src/types.ts
@@ -35,6 +35,7 @@ import * as StorefrontCaps from './filecoin/storefront.js'
 import * as AggregatorCaps from './filecoin/aggregator.js'
 import * as DealTrackerCaps from './filecoin/deal-tracker.js'
 import * as DealerCaps from './filecoin/dealer.js'
+import * as IndexCaps from './index/index.js'
 import * as AdminCaps from './admin.js'
 import * as UCANCaps from './ucan.js'
 import * as PlanCaps from './plan.js'
@@ -52,6 +53,8 @@ export interface UCANAwait<Selector extends string = string, Task = unknown> {
  * An IPLD Link that has the CAR codec code.
  */
 export type CARLink = Link<unknown, typeof CAR.codec.code>
+
+export type Multihash = Uint8Array
 
 export type AccountDID = DID<'mailto'>
 export type SpaceDID = DID<'key'>
@@ -449,6 +452,51 @@ export type UploadGetFailure = UploadNotFound | Ucanto.Failure
 // HTTP
 export type HTTPPut = InferInvokedCapability<typeof HTTPCaps.put>
 
+// Index
+export type Index = InferInvokedCapability<typeof IndexCaps.index>
+export type IndexAdd = InferInvokedCapability<typeof IndexCaps.add>
+
+export type IndexAddSuccess = Unit
+
+export type IndexAddFailure =
+  | IndexNotFound
+  | DecodeFailure
+  | UnknownFormat
+  | ShardNotFound
+  | SliceNotFound
+  | Failure
+
+/** An error occurred when decoding the data. */
+export interface DecodeFailure extends Failure {
+  name: 'DecodeFailure'
+}
+
+/** The data is not in a format understood by the service. */
+export interface UnknownFormat extends Failure {
+  name: 'UnknownFormat'
+}
+
+/** The index is not stored in the referenced space. */
+export interface IndexNotFound extends Failure {
+  name: 'IndexNotFound'
+  /** Multihash digest of the index that could not be found. */
+  digest: Multihash
+}
+
+/** A shard referenced by the index is not stored in the referenced space. */
+export interface ShardNotFound extends Failure {
+  name: 'ShardNotFound'
+  /** Multihash digest of the shard that could not be found. */
+  digest: Multihash
+}
+
+/** A slice referenced by the index was not found in the specified shard. */
+export interface SliceNotFound extends Failure {
+  name: 'SliceNotFound'
+  /** Multihash digest of the slice that could not be found. */
+  digest: Multihash
+}
+
 // Blob
 export type Blob = InferInvokedCapability<typeof BlobCaps.blob>
 export type BlobAdd = InferInvokedCapability<typeof BlobCaps.add>
@@ -458,9 +506,8 @@ export type ServiceBlob = InferInvokedCapability<typeof W3sBlobCaps.blob>
 export type BlobAllocate = InferInvokedCapability<typeof W3sBlobCaps.allocate>
 export type BlobAccept = InferInvokedCapability<typeof W3sBlobCaps.accept>
 
-export type BlobMultihash = Uint8Array
 export interface BlobModel {
-  digest: BlobMultihash
+  digest: Multihash
   size: number
 }
 
@@ -578,7 +625,7 @@ export interface StoreAddSuccessResult {
   /** DID of the space this item will be stored in. */
   with: DID
   /** CID of the item. */
-  link: UnknownLink
+  link: CARLink
 }
 
 export interface StoreAddSuccessDone extends StoreAddSuccessResult {
@@ -616,7 +663,7 @@ export interface ListResponse<R> {
 }
 
 export interface StoreListItem {
-  link: UnknownLink
+  link: CARLink
   size: number
   origin?: UnknownLink
   insertedAt: ISO8601Date
@@ -856,7 +903,9 @@ export type ServiceAbilityArray = [
   ServiceBlob['can'],
   BlobAllocate['can'],
   BlobAccept['can'],
-  HTTPPut['can']
+  HTTPPut['can'],
+  Index['can'],
+  IndexAdd['can']
 ]
 
 /**