+++ packages/filecoin-api/test/context/types.ts
@@ -7,6 +7,6 @@ export interface UpdatableStoreOptions<K, V> extends StoreOptions<K, V> {
   updateFn?: (items: Set<V>, key: K, item: Partial<V>) => V
 }
 
-export interface StreammableStoreOptions<K, V> extends StoreOptions<K, V> {
-  streamFn?: (items: Set<V>, item: K) => AsyncIterable<V> | undefined
+export interface ReadableStreamStoreOptions<K, V> extends StoreOptions<K, V> {
+  streamFn?: (items: Set<V>, item: K) => ReadableStream<V> | undefined
 }