+++ packages/upload-api/src/types.ts
@@ -15,6 +15,7 @@ import type {
   RevocationChecker,
   ToString,
   UnknownLink,
+  MultihashDigest,
   Unit,
 } from '@ucanto/interface'
 import type { ProviderInput, ConnectionView } from '@ucanto/server'
@@ -157,6 +158,9 @@ import {
   PlanCreateAdminSession,
   PlanCreateAdminSessionSuccess,
   PlanCreateAdminSessionFailure,
+  IndexAdd,
+  IndexAddSuccess,
+  IndexAddFailure,
 } from '@web3-storage/capabilities/types'
 import * as Capabilities from '@web3-storage/capabilities'
 import { RevocationsStorage } from './types/revocations.js'
@@ -191,6 +195,16 @@ import {
   BlobAddInput,
 } from './types/blob.js'
 export type { AllocationsStorage, BlobsStorage, TasksStorage, BlobAddInput }
+import { IPNIService, IndexServiceContext } from './types/index.js'
+export type {
+  IndexServiceContext,
+  IPNIService,
+  BlobRetriever,
+  BlobNotFound,
+  ShardedDAGIndex,
+  ShardDigest,
+  SliceDigest,
+} from './types/index.js'
 
 export interface Service extends StorefrontService, W3sService {
   blob: {
@@ -311,6 +325,9 @@ export interface Service extends StorefrontService, W3sService {
   }
   usage: {
     report: ServiceMethod<UsageReport, UsageReportSuccess, UsageReportFailure>
+  }
+  index: {
+    add: ServiceMethod<IndexAdd, IndexAddSuccess, IndexAddFailure>
   }
 }
 
@@ -465,6 +482,7 @@ export interface ServiceContext
     PlanServiceContext,
     UploadServiceContext,
     FilecoinServiceContext,
+    IndexServiceContext,
     UsageServiceContext {}
 
 export interface UcantoServerContext extends ServiceContext, RevocationChecker {
@@ -484,6 +502,10 @@ export interface UcantoServerTestContext
   fetch: typeof fetch
 
   grantAccess: (mail: { url: string | URL }) => Promise<void>
+
+  ipniService: IPNIService & {
+    query (digest: MultihashDigest): Promise<Result<Unit, RecordNotFound>>
+  }
 }
 
 export interface StoreTestContext {}
@@ -610,7 +632,7 @@ export type AdminUploadInspectResult = Result<
 
 export interface StoreAddInput {
   space: DID
-  link: UnknownLink
+  link: CARLink
   size: number
   origin?: UnknownLink
   issuer: DID