+++ packages/upload-api/test/handlers/index.spec.js
@@ -1,0 +1,4 @@
+import { test } from '../test.js'
+import * as Index from './index.js'
+
+test({ 'index/*': Index.test })