+++ packages/w3up-client/src/index.node.js
@@ -8,6 +8,7 @@ import { generate } from '@ucanto/principal/ed25519'
 import { Client } from './client.js'
 export * as Result from './result.js'
 export * as Account from './account.js'
+export * from './ability.js'
 
 /**
  * Create a new w3up client.