+++ packages/filecoin-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.3.3](https://github.com/w3s-project/w3up/compare/filecoin-client-v3.3.2...filecoin-client-v3.3.3) (2024-04-30)
+
+
+### Fixes
+
+* filecoin test use blob ([#1422](https://github.com/w3s-project/w3up/issues/1422)) ([359c0b7](https://github.com/w3s-project/w3up/commit/359c0b736cad8e4375d75af4f60e97e20057e7aa))
+
 ## [3.3.2](https://github.com/w3s-project/w3up/compare/filecoin-client-v3.3.1...filecoin-client-v3.3.2) (2024-04-24)